requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.0.0
elevenlabs>=0.2.0
pillow>=10.0.0
pdf2image>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
gTTS>=2.4.0
sentencepiece>=0.1.99
